requirements.txt
# network-delay-sensor dependencies
pyyaml
pycryptodome